更改dis squart

+++ include/autoManualFindRelation.hpp
@@ -83,8 +83,8 @@ public:
 	void calcNormalWay(Point2i inPoint,vector<Point2i>& triVertex, vector<Point2i>& triPos,Point2i& result);
 
 
-	float getDistance(Point2i pointO, Point2i pointA);
-	float getDist_P2L(Point2i pointP, Point2i pointA, Point2i pointB);
+	double getDistance(Point2i pointO, Point2i pointA);
+	double getDist_P2L(Point2i pointP, Point2i pointA, Point2i pointB);
 	void calcDistancePoint2Triangle(Point2i inPoint,vector<Point2i>& triVertex, vector<double>& dis);
 	void getNear2LineUseTwoPoint2Calc(int flag,Point2i inPoint,vector<Point2i>& triVertex, vector<Point2i>& triPos,Point2i& result);
